Changed coordinates to long

+++ src/filmparser/Coordinate.java
@@ -1,8 +1,8 @@
 package filmparser;
 
 public class Coordinate {
-    private int x;
-    private int y;
+    private long x;
+    private long y;
 
     @Override
     public String toString() {
@@ -18,19 +18,19 @@ public class Coordinate {
         this.y = y;
     }
 
-    public int getX() {
+    public long getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(long x) {
         this.x = x;
     }
 
-    public int getY() {
+    public long getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(long y) {
         this.y = y;
     }
 }
